+++ spec/factories/topic_section.rb
@@ -15,6 +15,7 @@ FactoryGirl.define do
     topic
     title "Topic Section Title"
     description "Topic Section Description"
+    position 0
 
     after(:build) do |topic_section, evaluator|
       evaluator.guides.each do |guide|
