+++ spec/models/topic_spec.rb
@@ -72,3 +72,32 @@ RSpec.describe Topic do
     end
   end
 end
+
+RSpec.describe Topic, '#add_section' do
+  it 'automatically sets position to 1 for the first section in a topic' do
+    topic = create(:topic)
+
+    topic_section = topic.add_section
+    expect(topic_section.position).to eq 1
+  end
+
+  it 'automatically sets position to the next section in an existing topic' do
+    topic = create(:topic)
+
+    create(:topic_section, position: 5, topic: topic)
+    topic_section = topic.add_section
+
+    expect(topic_section.position).to eq 6
+  end
+
+  it 'only considers sections in the same topic when setting position' do
+    topic = create(:topic)
+    another_topic = create(:topic, path: '/service-manual/another-topic')
+
+    create(:topic_section, position: 5, topic: another_topic)
+    another_topic.add_section
+    topic_section = topic.add_section
+
+    expect(topic_section.position).to eq 1
+  end
+end
